Cleaned up the includes in all of the files.
All of the includes are now in the .h files. The only includes in .cpp files should be the .h file for the class. the only exception for this is robot .h to include classes that need to include robot.h.

The includes have also been ordered to make them more readable. The order is as follows:
1. standard c++ includes (string, iostream, etc)
2. frc library includes (smartdashboard, phoenix, etc)
3. our includes, split up by type. (ex. opmodes, subsystems, etc)